Colin's changes to admin
Modified, edited, and cross pollinated:
- overview of spinnaker
- adding accounts

@imosquera for the Architecture page there was an incomplete sentence in
the opening paragraph - I deleted it because this is live, but feel free
to go back in and complete the thought.
the Target Cloud Providers has an incomplete sentence as well.

I'm going to assume you tried to define Stand alone and High
Availability at the bottom of Architecture, but it really breaks the
user flow. Can we talk about this? I'd move it elsewhere.