Document new installer (#349)
* Add "earlier version" warnings to all EC2 pages

Added placeholder "Install Armory Spinnaker" page for k8s installer as
target link.  Note the warning is an "included" component, so if the target
needs to change, just change the _includes/components/legacy_documentation.html
file.

* Started on the install instructions for Armory

Armory installer (for Mac OSX) installation, and instructions for running
the init subcommand to install Spinnaker.  TBD for Docker image.

* Adding --version latest to command line

This might not be needed if/when we make the 'stable' version available.